Extract Android code to header file
